Fix main component switching. Do not blow the dom.
The previous way would blow the whole DOM for `main-page` when what we really
wanted was that component to be updated.

Using this way with `case` will not blow the DOM when switching from :home and
:search and will also work with figwheel's live reload. You can change the
main-page component and save and it will properly update on the page without
blowing away your state.

Not sure why this way works though and the previous one didn't.